Bug in admin backend: backup/restore page. Fix attached in form of zipped patch file (as produced by Beyond Compare 3.x)
How To Reproduce / Observed Symptoms

in admin page: go to backup/restore page, window pops up, hit 'Delete' button in there without ticking ANY backup file and see a bare, 'unstyled' error message appear in the window.
The fix remedies this and 'styles' the error message like it should, same as other action feedback messages are styled in the popup window. The error report will be in a red box now.

The Code

backend.php has a 'weird' way to produce the HTML page: the HEAD section is only dumped halfway through, while various action sections above that bit may have produced various <p>/<div>y chunks of CONTENT, thus resulting in malformed HTML being fed to the client.

Oh, besides: the permissions are only checked AFTER the action has already happened -- the permission check only affects whether or not we get to hear about the results.

The Fix

The remedy is reshuffling the code in backend.php a bit; the HEADER section of the HTML is printed early, and so is the permission check. Both ensure actions only take place when permitted and any happy or sad reports of theirs end up in the designated slot in the HTML page, ensuring proper styling on any occasion.

See also:
http://community.compactcms.nl/forum/index.php/topic,178.msg681.html#msg681