* Bug in lightbox: when multiple albums are assigned to a single page, only the 'first' one shows up, ever     * Couple of oddities in the code fixed as well - call it cleanup (including encoding various POST/GET params on the way in)
How To Reproduce / Observed Symptoms

create a lightbox, create two (or more) albums and upload content to both; then assign both to the same (single) lightbox page in the CMS. See only one of 'em appear.
The fix remedies this and includes some code cleanups, but not all (the GET/POST/etc. XSS/SQL/etc. security fix done a-retentive style is reserved for another message at a later date.   :-\

The Code / The Fix

Side Note that lightbox.Manage.php doesn't encode the inputs, so raw form/URL access gets into the machine; okay, so it's the admin interface of CCMS, but nevertheless.
Manage was missing one 'echo' for cleanliness sake. Fixed.

Another side note shows lightbox.Process.php sporting a dangling 'else' which would 'consume' the next $error assignment. Lucky us everything after that is also part of that 'else' because of the die()ing code in the if() before. 'else' removed. Fixed.

Also note that the incoming content in encoded to ensure nothing untowards gets into the files/database. (I've not yet done a solid job on this, but we're only halfway re that.)

None of these concern the main problem: only one album shown on page, ever.

The actual fix about the listed problem is on lightbox.Show.php where we collect a matching ARRAY of albums and only set the 'single' flag when we find we've only got one at the end of that scan.
Add to that a filter within the display loop to only dump links for the matching == assigned albums and we got what we expected: an index of only the pages that were assigned to this particular lightbox page.