A little side note how to restore some code I just stripped out (not available in git anywhere!): I was completely stumped about why the captcha wasn't working like it should, so I added special code at the /*MARKER*/ places to track who included what, when.
And that was how I found out that under particular circumstances sitemap.php would, sensibly from its own perspective, load the current page, and that would mean the plugin PHP files. But those would fail to load. ... Because sitemap.php would be require_once()'d by a plugin's Process.php, which would be invoked by a form submission or AJAX action, which, of course, wouldn't stick to the one-two level deep structure as presented to the outside by CCMS for regular URLs.