lightbox module: fixed bugs in image+thumbnail scaling (producing huge images while size seems restricted to 640x480 in the code; the lightbox kinda needs this) & guaranteed display sequence.
The image scaling logic was completely buggy; it failed for any image with an aspect ratio larger than the desired 640:480 for the 'full-sized' image show and for any image which didn't have the 1:1 (80x80) aspect ratio for thumbnails.

Fixed by:

1) adding correct rescaling logic for both the 'full-size' and thumbnail calculations fed into GD

The thumbnails did keep their original aspect ratio at least on disc, but not in the thumbnail view where any image would be stretched to exactly fit the 80x80px box.

2) And that was fixed by adding additional padding calculation logic to ensure the right amount of padding of the thumbnail is specified for any thumbnail image shown.

----

The directory scanning logic uses opendir/readdir and, at least on network drives, this doesn't ensure you'll get a deterministic order of the images in the directory.
As the first image in the set is used as the thumbnail view for the overview page, this is relevant in more than one way.

The new code scans the album directory and then sorts the set on filename to ensure a deterministic sequence is produced, always.