Significantly reduce the number of queries in the 'edit ownership' screen: there, more than USERS x PAGES queries would be executed upon hitting SAVE as each user is seperately added to each page's ownership list.
The new code collects the users in a fast PHP array and then updates each page exactly once.

The new code also does NOT reset all pages ownership to 'NIL' beforehand, which matters when we add a feature which filters/reduces the number of pages we manage, so full- and reduced-rights users can filter the pages they wish to see/edit/manage.