fileperms() --> is_writable() + move so read-only files on Windows servers are detected correctly (weren't so upto now).
TODO: get rid of the UNIXy fileperms() == 0666 checks everywhere; it's already happened in the template editor (admin screen) but the installer section still needs to be 'processed' like that.

This is just filing the TODO in the source code.

ADDED FEATURE: templates aren't just a HTML plus a CSS, so the 'template editor' now recognizes the usual text files which may part of a template: CSS, JS, more HTML (template variant for printing!) and PHP (as the template may decide to require/include other PHP files on its own...)